clapper: glimmer: fix for keyboard occasionally disabled on wake
Fixes bug in which the keyboard remains disabled after resuming.
The problem occurs when the lidangle_keyscan_update() occurs after
the CHIPSET_RESUME hook triggers and before chipset state is
CHIPSET_STATE_ON, and the lid angle is in range to disable the
keyboard.

The fix is to not use a hook at all on chipset resume, and instead
just keep track of the last chipset state locally in
lidangle_keyscan_update(). That way we guarantee that when the
AP is turned on, the next time we check for lidangle keyscanning
state, it will enable the keyboard.

BUG=chrome-os-partner:28480
BRANCH=rambi
TEST=tested on a glimmer:

Added a hook to detect chipset resume in lid_angle.c:
static void enable_keyboard(void)
{
	hook_fired = 1;
}
DECLARE_HOOK(HOOK_CHIPSET_RESUME, enable_keyboard, HOOK_PRIO_DEFAULT);

Then in lidangle_keyscan_update() added:
if (hook_fired && !chipset_in_state(CHIPSET_STATE_ON))
        ccprintf("Hook fired, but chipset state still reads off\n");
hook_fired = 0;

Using this code, I used the lid switch on the servo board to open
and close the lid repeatedly until I saw the printf above.

Without this CL, when the printf is shown, the keyboard remains
disabled in S0. With this CL, the keyboard is enabled in S0.

Change-Id: Iac017a6468d46e0f011ec9d84a8c73b89622d8cc
Signed-off-by: Alec Berg <alecaberg@chromium.org>