clapper: glimmer: fix for keyboard occasionally disabled on wake
Fixes bug in which the keyboard remains disabled after resuming.
The problem occurs when the lidangle_keyscan_update() occurs after
the CHIPSET_RESUME hook triggers and before chipset state is
CHIPSET_STATE_ON, and the lid angle is in range to disable the
keyboard.

The fix is to not use a hook at all on chipset resume, and instead
just keep track of the last chipset state locally in
lidangle_keyscan_update(). That way we guarantee that when the
AP is turned on, the next time we check for lidangle keyscanning
state, it will enable the keyboard.

BUG=chrome-os-partner:28480
BRANCH=rambi
TEST=none

Change-Id: Iac017a6468d46e0f011ec9d84a8c73b89622d8cc
Signed-off-by: Alec Berg <alecaberg@chromium.org>