Add info_printf() and ccinfo_printf()
Our code base contains a lot of debug messages in this pattern:
  CPRINTF("[%T xxx]\n") or ccprintf("[%T xxx]\n")
The strings are taking up spaces in the EC binaries, so let's refactor
this by adding info_printf() and ccinfo_printf().

info_printf() is just like cprintf(), except that it adds the brackets,
the timestamp, and the name of the console channel. ccinfo_printf() is
equivalent to info_printf(CC_CONSOLE, ...).

This saves us hundreds of bytes in EC binaries.

BUG=chromium:374575
TEST=Build and check flash size
BRANCH=None

Change-Id: Ifafe8dc1b80e698b28ed42b70518c7917b49ee51
Signed-off-by: Vic Yang <victoryang@chromium.org>