Update patch set 1
Patch Set 1: Code-Review-2

I'm not going to approve this change. Chrome already re-maps  scancodes to keycodes for different language layouts. If we are going to put an exception in to re-map the R_CTRL key for kip with the FCE layout it should be done in the same place as the other re-maps.

Patch-set: 1
Label: Code-Review=-2