Increase UART_RX_DMA_RECHECKS
This increases the number of times the UART is checked for input
when an RX interrupt is received in the DMA case. The current value
of 5 is good for interactive sessions, but we also need it to
accomodate time-sensitive commands from automated tests so that
things like "reboot ap-off" take effect before the vboot on the
AP-side has a chance to run and potentially change the state.

BUG=none yet
BRANCH=nyan
TEST=firmware_RecoveryButton FAFT test succeeds more reliably

Change-Id: I4618473c3b56c2a56fa9a9632a9afdb481d16365
Signed-off-by: David Hendricks <dhendrix@chromium.org>