fruitpie: new PD TX driving for reworked board
For boards reworked with FETs on the TX path, TX_DATA (PB14) needs to be
driven low when we are not transmitting.
Set MISO as GPIO low by default and set the alternate SPI function in
the TX enable code.

This should be backward compatible with non-reworked Fruitpies.

Signed-off-by: Vincent Palatin <vpalatin@chromium.org>

BRANCH=none
BUG=none
TEST=plug 2 fruitpies with a type-C cable and do a PD power
negotiation.

Change-Id: I9ef9c84147e9468667eed9c96647c6a9a7f79fdd