build: Move libs after source or object files
When not using gold as the linker the ordering of libraries
with respect to objects and source files is important.
Previously the build placed -l libraries before source and
objects on the gcc command line.  This doesn't work with
the default ld linker because since none of the symbols in
the libraries are yet required they are thrown out before
their uses are found in the objects and source.

BUG=none
BRANCH=none
TEST=make buildall

Change-Id: Ic9f83ba6138d6592d3b6e28de6fb0688e664f480