Update patch set 1
Patch Set 1:

This is a good start. It seems like it might also be useful to have an option to set this bringup flag for a single boot without have to compile a new binary. Like how we have the EC console command "reboot ap-off", it could be a "reboot bringup" command.

Patch-set: 1