battery: samus: allow charging of dead battery reporting 0's
On samus battery, when the battery is dead it reports 0 for desired
voltage, current, and state of charge. In this case we should allow
charging.

With this CL, when a dead samus battery is used and a charger is
connected, we attempt to charge it, and get a bunch of i2c bus errors
for a little while, presumably because the battery MCU is in some
weird state. After ~10 seconds though, the battery stabilizes and
we go through the precharge and then charge as normal.

BUG=none
BRANCH=none
TEST=test on a samus with a dead battery. w/o this CL, the battery
never charges because the charging not allowed flag is set. With this
CL, there is a rocky period of i2c errors at first, but then it
charges normally.

Change-Id: Ic61f27a27237166d33cb9ea5f024d3ef6360ce82
Signed-off-by: Alec Berg <alecaberg@chromium.org>