charger v2: set charger mode when requesting current/voltage
Clear the CHARGE_INHIBIT bit when requesting non-zero current/voltage
and set it when charging should be disabled.  On Blaze with the
BQ24725 charger, setting this bit saves 10-15mW when we're not on AC.

This also fixes an issue where battery charging would not get enabled
when the charger is connected while the machine is in S3/S5.  This is
because the kernel driver would inhibit charging when the charger was
removed and then the EC would not enable it when the charger was
re-connected while the host was off, such as in S3/S5.

BRANCH=nyan
BUG=chrome-os-partner:29386
TEST=Boot Blaze, disconnect charger, suspend, connect charger and
observe that the battery now starts charging.

Change-Id: I3a8f30a307391248b6b8b005ed31c6c1bc847532
Signed-off-by: Andrew Bresticker <abrestic@chromium.org>