IRQ list support for enabling specific IRQs without common runtime
This is a preparatory work for adding back DECLARE_IRQ() support when
building without common runtime. With this, we can enable only a subset
of IRQs and avoid linking in other unused IRQ handlers.

Note that after this change, all boards without common runtime need to
have a ec.irqlist file.

BUG=None
TEST=Build Keyborg and check it still works.
TEST=make buildall
BRANCH=None

Change-Id: If68062a803b9a78f383027a1625cf99eb3370d3f
Signed-off-by: Vic Yang <victoryang@chromium.org>