Update patch set 5
Patch Set 5: Verified+1

(1 comment)

The conversion process was partially automatic and partially manual.  I directly copied the contents of the board.h and board.c files into the gpio.inc file.  Then I automated the interleaving of the lines.  Than manual process was comparing each pair of lines to verify that the enum name and gpio string name matched.  So I have a high confidence that the conversion is correct.  There were a couple of location where the names in the enum and the string name of the GPIO were different.  After talking with Randal and Bill all of the string names were brought in line with the enum names.

Unless there is an objection I'll land this change and make the >80-column change to the .inc files next.

Patch-set: 5
Label: Verified=+1