zinger: update unhandled exception panic
When getting an unhandled exception,
print a trace with the PC and LR if the debug output is activated
also properly reboot the CPU rather than just jumping to the reset
vector as the previous code (that might lead to some incorrect
initializations).

Signed-off-by: Vincent Palatin <vpalatin@chromium.org>

BRANCH=none
BUG=chrome-os-partner:29840
TEST=add adhoc code triggering a data abort and see the firmware
printing a trace, then rebooting immediatly in a working state.

Change-Id: I1d5a98d9113c8ae08e05588a40f941d1ed22cebe