cprints: Revert some changes from CPRINTF to CPRINTS
These changes were made in files that did not have the [%T ... ]
pattern.  These files were broken by the change because they still
contained uses of the CPRINTF macro.  There were two options to fix
this, switch to the CPRINTS macro and get the timestamp added to
these strings, or switch those files back to defining the CPRINTF
macro.  Switching back seems like the right thing since it doesn't
change the output of those debug messages.

This commit also adds newline termination to a few invocations of
CPRINTF that were missing it, but obviously wanted it.

This breakage is only visible with a particular set of CONFIG_
defines that no boards currently use.

Signed-off-by: Anton Staaf <robotboy@chromium.org>

BRANCH=none
TEST=make buildall -j

Change-Id: I784b52dc385b29f05d7b9bc1521e37597409153b