ryu: add power sequencing
Update the Ryu board configuration to re-use the Tegra power state
machine.
We have no dedicated HOLD GPIO on this board to detect when the AP wants
to shutdown, for now I'm using the EC interrupt net (which has a pull-up
on the AP side) as an approximation.

Signed-off-by: Vincent Palatin <vpalatin@chromium.org>

BRANCH=none
BUG=none
TEST=on Ryu, use the power button to switch on and off the AP.

Change-Id: I310438e0b923956d1539fb919c266a94909e3461