Update patch set 16
Patch Set 16: Code-Review-1

(3 comments)

The addition of last_tx_ok seems odd, other than that just a few nits.  Thanks for taking the time to clean this up before it lands.

As for moving the check into the __tx_char function, now that is_reset is checked in usb_wait_console, which is called by usb_putc, we're covered.

Patch-set: 16
Label: Code-Review=-1