[HACK] Console command to get the battery out of cut-off
Somehow the batteries go into cut-off mode sometimes. We should be able
to revive the battery by plugging in AC power, but that's not working
now. Until then, let's add a console command for this. The command sets
the charger to supply voltage 0.75V above the battery voltage for 50 ms
so as to kick the battery out of the cut-off mode.

BUG=chrome-os-partner:30633
TEST=Kill a battery by shorting cut-off signal to ground. Revive it with
the new command.
BRANCH=None

Change-Id: I1cedffbf6ae6afea2fd0d00bbfddf931f71e124b
Signed-off-by: Vic Yang <victoryang@chromium.org>