pd: sink change sending getSourceCap to hard reset after timeout
Changes to match this part of the spec:

The Sink shall support the SinkWaitCapTimer. When a Sink observes an
absence of Source Capabilities messages, after VBUS is present, for a
duration of tTypeCSinkWaitCap for the Type-C connector and
tSinkWaitCap after VBUS is present,for all other connectors the Sink
shall issue Hard Reset signaling in order to restart the sending of
Source Capabilities messages by the Source (see Section 6.6.4).

tTypeCSinkWaitCap shall be between 210 and 250 ms

Also set send source cap period appropriately:
tTypeCSendSourceCap shall be between 100 and 200 ms

This should help avoid transmission collisions during negotiation.

BUG=chrome-os-partner:30135
BRANCH=none
TEST=load onto zinger and samus and verify they negotiate correctly
10 times. Also loaded custom code to zinger to not send source cap
and verified we send hard reset.

Change-Id: Idd6118e3e0a9f7a96ebeae9518c8b10457232c70
Signed-off-by: Alec Berg <alecaberg@chromium.org>