FIFO: Add byte FIFO
This is very similar to the existing queue and should be merged.
That merge requires some discussion and thought though.  So I'm
putting it off so we can talk about the USART driver for now.  I
don't expect this commit to go into the tree.  I expect it to be
replaced with a commit that merges this functionality into the
existing queue code and an update to the later commits to use the
queue instead of this FIFO.

Change-Id: Ia560d6ef2344f8044c7fb25bfef575a6847f5d23