stm32-USART: Add generic stream based usart driver
This driver can be used to access multiple usarts using an abstract
stream interface.  The stream interface can also be used in drivers
for the host interface and USB console interface, providing a
consistent API across all character stream style IO.

This commit isn't finished, I wanted to get it up so we can start
talking about it now that I think it is in the right shape.  It works,
but has some rough edges still.

Change-Id: Icf567f0b0fa4eb0e9ad4cdb0be8edc31c937a7de