Remove dependency of flash_ec on crosutils
The script currently depends on common.sh from crosutils, which is not
installed on beaglebone and also pulls in dependency on other
repositories. As we are only using shflags and output formatting from
common.sh, let's put those libraries in EC repo so as to remove the
dependency on crosutils.

BUG=chromium:397202
TEST=Run the script to flash EC
BRANCH=None

Change-Id: Ib18987410eb32d773d55fb4e53133adf230167b9
Signed-off-by: Vic Yang <victoryang@chromium.org>