Update patch set 1
Patch Set 1:

Testing shows that turning off the 5V regulator is more effective at catching the fault. I'm also turning off USB power for good measure. The existing fix for Peppy seems to work with this "cable of consternation" as is so I don't plan on pushing a Peppy EC firmware for this.

Patch-set: 1