Update patch set 3
Patch Set 3:

Expanding the eyes on this work, since it is a major refactor of how we interface to the UART.  The plan is to write in/out_stream drivers for the AP I2C interface to the console, as well as to the USB bulk endpoint serial interface and use the generic stream interface to mux everything together.  This will eventually lead to a clean up of the uart_buffering code as well.

Patch-set: 3
Label: Code-Review=0