Update patch set 1
Patch Set 1: Code-Review-1

I'm not convinced that's better :
if we are sure to always trigger reception on the beginning of the message : that's great.
but we triggered too early, we are going to cut the message in the middle due to the timeout.

if we want to put a very tight timeout, we would need to check it with more complex conditions , but i did not do it because that was sucking a lot of CPU power at RX.

Patch-set: 1
Label: Code-Review=-1