Update patch set 1
Patch Set 1:

Ok, I agree. We are going to need to look at this closer because right now, we wait for three retries with this blocking delay of 2.7ms, which causes problems for the lower priority port. We may need to setup an interrupt for the end of reception, for example the CC line going back to it's idle/connected voltage.

Patch-set: 1
Reviewer: Gerrit User 1103909 <1103909@3ce6091f-6c88-37e8-8c75-72f92ae8dfba>
Label: Code-Review=0