Update patch set 4
Patch Set 4:

@Mohammed Habibulla		Aug 2 12:57 PM

charger_closest_voltage checks the same condition. How is this different?

Also please list on the issue tracker the exact steps which cause the issue that you are trying to fix.
		
@YongBeum Ha		Aug 2 1:50 PM

The parameter of charger_closest_voltage() is batt->desired_voltage and the value that is compared here is batt->volatage.

charger_closest_voltage() is to get voltage that is set to charger. This line is to check battery voltage itself. If battery voltage is higher than maximum voltage, we think that the battery has defect and don't charge the battery permanently.
	


@Mohammed Habibulla		Aug 2 12:57 PM

Once the error occurs you need to disconnect and reconnect the battery to clear the error?

@YongBeum Ha		Aug 2 1:50 PM

Yes. If battery has defect, we don't charge the battery. Disconnection and reconnection can be inserted new battery, we must clear the error.

Patch-set: 4
Label: Code-Review=0