Update patch set 5
Patch Set 5:

(5 comments)

Thanks everyone.  I'm working on the flushing semantics, which include the possibility of interrupt handler re-entrance.  I understand that the hardware won't re-enter the interrupt, but it is possible (I believe, not positive) to get a re-entrant condition as the flushing code stands.  I hope to make that impossible.

Patch-set: 5