Update patch set 1
Patch Set 1:

I suppose we can hold off on this for now, since...

Samus doesn't need this change (when the LM4 hibernates to hard-reset, that resets PD)

Ryu doesn't have a separate PD chip

Everything farther off is still in flux.  We may need to revisit this if we have a MEC1322-based EC + a separate STM32 PD chip, or a STM32+STM32.  But hopefully discrete PD chips will appear by then.

Patch-set: 1