Update patch set 3
Patch Set 3:

(5 comments)

Testing is a little tricky right now because of the hardware bugs on samus. If you refer to the "Samus" section in the "Closed-Case Debugging: Details and Status" document, the Silego chip needs to be replaced (U17), and associated FETs reversed (Q3, Q5, Q16). There is also a conflict between USB PD C1 and SPI flash resources over SPI2 and DMA channels 6/7, so for testing I have a temporary CL to disable PD. After that, there is another hardware bug with the SPI flash not being powered, so the servo cable needs to be connected and "dut-control spi2_vref:pp3300" sent. Then, from the console type "debugset 1", and the "spi_flashinfo" command should execute successfully.

For testing USB mass storage, after fixing SPI flash as above, I connect the board to the computer, and the SPI flash should be read successfully as a block device in Linux. The "sdparm -i /dev/sdX" and "sg_inq /dev/sdX" should also complete successfully and return valid data. For verifying data integrity, refer to the "test_spi_integrity.sh" script in the shared folder, which repeatedly takes an image of the SPI flash, writes it into /dev/sdb, reads it back, and checks to make sure the two are the same. Make sure to read through any of the scripts before executing them, they are hardcoded and may cause problems if you have an actual disk at /dev/sdb.

Patch-set: 3
Label: Code-Review=0